% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot3.R
\name{myboot3}
\alias{myboot3}
\title{myboot3}
\usage{
myboot3(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of times to bootstrap}

\item{x}{Data used}

\item{fun}{Function to use, mean, median etc}

\item{alpha}{Determines confidence interval}

\item{cx}{Plotting parameter}

\item{...}{used for histogram}
}
\value{
None (variables suppressed)
}
\description{
Basic bootstrapping function from Lab 9
}
\examples{
myboot3(10000,x=c(1,2,3),fun="var",alpha=0.05,xlab="var",col="Purple",cx=1.5)



}
