% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt2.R
\name{myclt2}
\alias{myclt2}
\title{myclt2}
\usage{
myclt2(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{number to sample}

\item{iter}{iterations}

\item{a}{parameter for uniform distribution}

\item{b}{parameter for uniform distribution}
}
\value{
None.
}
\description{
Samples from a uniform distribution.
}
\examples{
myclt2(n=10, iter=10000)


}
