% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm2.R
\name{mymlnorm2}
\alias{mymlnorm2}
\title{mymlnorm2}
\usage{
mymlnorm2(x, mu, sig, ...)
}
\arguments{
\item{x}{Values of x}

\item{mu}{mu}

\item{sig}{significance}

\item{...}{used for histogram}
}
\value{
List with x,coord,maxl
}
\description{
Maximum likelihood function from Lab 10
}
\examples{
mymlnorm2(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)



}
