% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(n, gamma, p)
}
\arguments{
\item{n}{The number of seats}

\item{gamma}{Chance of overbooking}

\item{p}{Chance of passengers taking seats}
}
\value{
List of n, p, gamma, and calculated values with continuous and discrete distributions.
}
\description{
Calculates number of airline tickets to sell using continuous and discrete distributions.
Minimizes the absolute value
}
\examples{
ntickets(200, .02, .95)


}
